+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 -----
 
 *   New Features:
-    *   Fixed background color for screens using the compose theme
-        ([#432](https://github.com/Automattic/pocket-casts-android/pull/432)).
+    *   Added Tasker integration with "Play Filter" and "Control Playback" actions.
+        ([#415](https://github.com/Automattic/pocket-casts-android/pull/431)).
+    *   Import OPML from a URL
+        ([#482](https://github.com/Automattic/pocket-casts-android/pull/482)).
+    *   Redesign of the fullscreen player share option
+        ([#451](https://github.com/Automattic/pocket-casts-android/pull/451)).
 *   Bug Fixes:
     *   Fixed Help & Feedback buttons being hidden when using text zoom.
         ([#446](https://github.com/Automattic/pocket-casts-android/pull/446)).
     *   Fixed when system bar didn't disappear on full screen video player
         ([#461](https://github.com/Automattic/pocket-casts-android/pull/461)).
-    *   Fixed when full screen video player was not closing on the first time in landscape mode
+    *   Fixed background color for screens using the compose theme
+        ([#432](https://github.com/Automattic/pocket-casts-android/pull/432)).
+    *   Fixed full screen video player not closing the first time in landscape mode
         ([#464](https://github.com/Automattic/pocket-casts-android/pull/464)).
 
 7.25